// Moviein/servidor/src/models/DTOs/RegisterUserDTO_Req.ts
interface RegisterUserDTO_Req {
    bairro: string
    cep: string
    cidade: string
    complemento: string
    confirmarSenha: string
    cpf: string
    dataNascimento: string
    email: string
    estado: string
    etapa: number
    genero: string
    nomeCompleto: string
    nomeMaterno: string
    numero: string
    pais: string
    senha: string
    telefone: string
}

export default RegisterUserDTO_Req;